requirements.txt
# Data Handling & Analysis
pandas-3.0.1
numpy-2.4.3

# Machine Learning
scikit-learn-1.8.0

# Data Visualization (For EDA before PowerBI)
matplotlib-3.10.8
seaborn-0.13.2

# Notebooks for testing & building
jupyter-1.1.1

# Web Application Backend (For webapp.py)
flask-3.1.3
flask-cors-6.0.2

# Database for storing user data & model outputs
pymongo-4.16.0

# Task scheduling for regular model retraining
schedule-1.2.2
